requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.1
django-phonenumber-field==5.0.0
phonenumberslite==8.12.8
pytz==2020.1
sqlparse==0.3.1